+++ lustre/llite/llite_lib.c
@@ -2527,9 +2527,9 @@ int ll_process_config(struct lustre_cfg *lcfg)
 
 /* this function prepares md_op_data hint for passing ot down to MD stack. */
 struct md_op_data * ll_prep_md_op_data(struct md_op_data *op_data,
-                                       struct inode *i1, struct inode *i2,
-                                       const char *name, int namelen,
-                                       int mode, __u32 opc, void *data)
+				       struct inode *i1, struct inode *i2,
+				       const char *name, size_t namelen,
+				       __u32 mode, __u32 opc, void *data)
 {
         LASSERT(i1 != NULL);
